breaking: roles.ts renamed to eicrud.roles.ts + default export inputDir set to ./src (not ./src/sevices)
