+++ cli/templates/eicrud-roles.ts
