+++ docs/security/roles.md
@@ -1,6 +1,6 @@
 Every ability in Eicrud is attached to a role. Roles are dynamically assigned to users via the property [CrudUser](../user/definition.md)->`role`.
 
-```typescript title="roles.ts"
+```typescript title="eicrud.roles.ts"
 import { CrudRole } from "@eicrud/core/config";
 
 export const roles: CrudRole[] = [
